Refactor system for multi platform support
In order to handle multiple frameworks we need to extract the framework
dependencies. At the same time its best to extract the parsing functions
to separate spi and serial methods as they are different in a few ways.

This will prepare the system for Zypher, IDF and Mbed. This will also
enable SPI drivers.