WIP: core driver refactor
This is a whole system refactor to make multi protocol and multi
platform support easier. Extract platform dependencies to lowest class,
extract protocol dependencies to middle class. This will also add
experimental versions of the ArduinoNativeAnt, MbedSerialAnt and
MbedSpiAnt classes. The ArduioAnt class is being renamed to the
ArduinoSerialAnt class to make it clearer the underlying protocol. There
is an opportunity to add a ArduinoSpiAnt but the protocol would have to
be hacked into a software based spi class, but one problem at a time,
and to be honest, if you are using the radio as a spi slave... get off
arduino.

This commit is a WIP and unstable. The only reason it is getting pushed
is beause I need to test on CI.