Native: Implement backfill model
This shoudl behave almost exactly the same as the serial model, the only
differences will be that messages cannot be buffered up. After every
call the response must be purged fully otherwise you risk overwriting
your previous request.

NOTE: this is still a WIP and does not compile but is pushed for code
sharing purposes